backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
redis==5.2.1
python-multipart==0.0.20

# Celery & Task Queue
celery==5.3.6

# Git Integration
gitpython==3.1.40

# AI & LangGraph
anthropic==0.47.0
langgraph==0.2.59
langchain-core==0.3.41
langchain-anthropic==0.3.9

# AWS Integration
boto3==1.35.86

# Security & Auth
cryptography==42.0.0
requests==2.31.0
